agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = ""
authors = [{ name = "Meng Yan", email = "myan@redhat.com" }]
# readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
  "ag-ui-protocol",
  "openai",
  "fastapi",
  "uvicorn",
  "python-dotenv",
  "copilotkit",
  "langgraph>=0.2.62,<0.3.0",
  "langchain",
  "langchain-openai",
  "langchain-mcp-adapters",
  "langchain-ollama",
  "nest-asyncio",
  "ipykernel>=6.30.1",
  "jupyter>=1.1.1",
]

[project.optional-dependencies]
dev = ["pytest", "ipykernel", "graphviz"]

[project.scripts]
main = "agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]
